Deep audit: fix 25 verified bugs across editor, runtime, and compat
Seven-subsystem audit (map editing, app shell, database, events, runtime
core, scenes/compat, forge). Every fix below was verified against the
code before changing it; runtime fixes verified by booting the Demo
template into a live battle (probe asserts sibling order + no errors);
editor fixes by 141/141 npm suite, 33/33 headless MapEditor tests, and
the NW.js map-load/undo harnesses.

Data safety: atomic project writes (temp+rename) everywhere; deploy
saves before building; dist worker rebuilds archives from scratch.

Database: autotile flag edits indexed kind*48 + mirrored across 48
shapes (never worked); class param curves 1-based (read one level low);
enemy HP/MP conditions parseFloat (collapsed to 0 on OK); Attack
Element off-by-one; Cancel baseline no longer re-taken per category.

Events: 402/403/411 markers at header indent (branches authored in
Reactor misrouted at runtime); all 97 insertion sites rebase indents to
the insertion context; block-header delete takes the whole block;
Change Gold param order; variable Transfer Player preserved on edit;
event click no longer wipes redo.

Map editor: targeted listener removal (blanket off() killed panning);
panning implemented via container movement; wheel/resize/scrollbar
listeners deduped (zoom compounded per map load); region overlay
survives map switches; Regions-tab eraser erases regions; moveMapBefore
assigns order; undo snapshots via slice() (3x whole-map JSON per stroke
gone); fill drops throwaway mid-flood shape calc; preview early-outs on
same tile pair; checkerboard RenderTexture cached.

Runtime: MZ battle sibling order restored under the Tier-1 MV
battle-field shim (battlebacks drew over battlers; boot-probe verified
1->4 above backs); duplicate Spriteset_Map.destroy merged (balloon
soft-lock); MZ event collision rule restored (MV rule to Tier 2);
Sprite/TilingSprite replaced-texture leak and tilemap _v8TexCache leak
closed (flagged-ownership destroy); WebAudio destroy leaves the load
watchdog (zombie re-downloads); watchdog sweeps throttled (O(N^2) at
load gates); _refresh allocation-free on unchanged frames; JsonEx no
longer serializes the MV _character mirror into saves (post-load wait
freeze); Decrypter encryption flags are live getters; Geometry
addAttribute shim gated to v8 (corrupted v5-v7 vertex data).

Forge: AnimationGenerator keyframes survive tool mounts (loader dropped
them; autosave made it permanent); Reset Layer keeps the keyframe
alias; pending autosave cancelled on detach (cross-project overwrite).

Deferred verified findings: docs/AUDIT-BACKLOG-2026-07-13.md.
Runtime synced to all four templates and Project1.

+++ editor/src/database/DatabaseEnemyEditor.js
@@ -663,8 +663,10 @@ class DatabaseEnemyEditor {
         okBtn.addEventListener('click', () => {
             action.skillId = parseInt(document.getElementById('action-edit-skill').value) || 1;
             action.conditionType = parseInt(document.getElementById('action-edit-condType').value) || 0;
-            action.conditionParam1 = parseInt(document.getElementById('action-edit-param1').value) || 0;
-            action.conditionParam2 = parseInt(document.getElementById('action-edit-param2').value) || 0;
+            // HP/MP/state-chance condition params are fractional rates
+            // (0.3 = 30%); parseInt collapsed them to 0 on every OK click.
+            action.conditionParam1 = parseFloat(document.getElementById('action-edit-param1').value) || 0;
+            action.conditionParam2 = parseFloat(document.getElementById('action-edit-param2').value) || 0;
             action.rating = Math.max(1, Math.min(10, parseInt(document.getElementById('action-edit-rating').value) || 5));
 
             enemy.actions[actionIndex] = action;
